refactor: 패키지 구조 변경
패키지 구조 변경 및 DrawRule 로직 변경
