--- src/main/java/nextstep/ladder/entity/AlwaysDrawRule.java
@@ -1,8 +1,0 @@
-package nextstep.ladder.entity;
-
-public class AlwaysDrawRule implements DrawRule {
-    @Override
-    public Direction nextDirection() {
-        return Direction.RIGHT ;
-    }
-}
