--- src/main/java/nextstep/ladder/entity/DrawRule.java
@@ -1,6 +1,0 @@
-package nextstep.ladder.entity;
-
-@FunctionalInterface
-public interface DrawRule {
-    Direction nextDirection();
-}
