--- src/main/java/nextstep/ladder/entity/LastPointNoneDrawRule.java
@@ -1,8 +1,0 @@
-package nextstep.ladder.entity;
-
-public class LastPointNoneDrawRule implements DrawRule {
-    @Override
-    public Direction nextDirection() {
-        return Direction.NONE ;
-    }
-}
