+++ src/main/java/nextstep/ladder/entity/draw/AlwaysDrawRule.java
@@ -1,0 +1,7 @@
+package nextstep.ladder.entity.draw;
+
+public class AlwaysDrawRule implements DrawRule {
+    public boolean drawRule() {
+        return true ;
+    }
+}
