+++ src/main/java/nextstep/ladder/entity/draw/DrawRule.java
@@ -1,0 +1,6 @@
+package nextstep.ladder.entity.draw;
+
+@FunctionalInterface
+public interface DrawRule {
+    boolean drawRule();
+}
