+++ src/main/java/nextstep/ladder/entity/draw/RandomDrawRule.java
@@ -1,0 +1,10 @@
+package nextstep.ladder.entity.draw;
+
+import java.util.Random;
+
+public class RandomDrawRule implements DrawRule {
+    @Override
+    public boolean drawRule() {
+        return new Random().nextBoolean();
+    }
+}
