+++ src/main/java/nextstep/ladder/entity/ladder/NoneDrawRule.java
@@ -1,0 +1,10 @@
+package nextstep.ladder.entity.ladder;
+
+import nextstep.ladder.entity.draw.DrawRule;
+
+public class NoneDrawRule implements DrawRule {
+    @Override
+    public boolean drawRule() {
+        return false;
+    }
+}
